Overview page has been re-done.
It now displays only a list of recent incidents. For each incident, it
displays the number of new updates since the user last logged out, and
shows links to the most recent of those updates.

For users with permission to approve users, it now also shows a
notification when there are users pending.