Created a new homepage for Collabbit.
This is the first revision. Copy and images need updating, not all links work, and other pages on the /home controller are probably now styled incorrectly.

This should be refactored out of Collabbit relatively soon. It's the Collabbit website, not part of the app. It's here now so we can get up and running quickly.