seeds.rb now generates random people for demo data
About ~60 random people are generated. All are using email addresses
@example.com, and all of their phone numbers are in the 555-01xx range,
which is reserved for fictitious use. They're generated from a list of
common first and last names. The password for all of them is 'demo'.

The test users have also been redone. They're now "John Super", "Jane
Admin", and "Joey User" and are a super admin/admin/normal user. Their
emails are test1/2/3@collabbit.org, and the password for all of them is
'demo'.