Improved display of lists throughout Collabbit.
Before, many of the lists were styled using the 'listy' class and a
sitewide style. There were several problems with that style and it
wasn't consistent with the rest of Collabbit. It has since been removed,
and the pages for group types, groups, and incidents have been re-styled
to better fit the Collabbit theme.